Use alternate database for travis ci builds

+++ tests/testthat/test_init.R
@@ -2,22 +2,24 @@ context('DB')
 library(nysatmoschem)
 
 ## create temp database
-dbname = basename(tempfile(pattern=''))
-if ('TRAVIS' %in% names(Sys.getenv())) {
-  db_res = system2('createdb', dbname, env = 'PGPORT=5433')
-} else {
-  db_res = try(system2('createdb', dbname))
-}
-if (is(db_res, 'try-error')) {
-  has_temp_db = FALSE
-} else {
+travis_ci = 'TRAVIS' %in% names(Sys.getenv())
+if (travis_ci) {
+  dbname = 'travis_ci_test'
   has_temp_db = TRUE
-  dbcon = src_postgres(dbname = dbname)
+} else {
+  dbname = basename(tempfile(pattern=''))
+  db_res = system2('createdb', dbname)
+  if (db_res == 0) {
+    has_temp_db = TRUE
+    dbcon = src_postgres(dbname = dbname)
+  } else {
+    has_temp_db = FALSE
+  }
 }
 
 ## make sure the temporary database is dropped even if there's an
 ## error
-if (has_temp_db) {
+if (!travis_ci && has_temp_db) {
   on.exit({
     try(DBI::dbDisconnect(dbcon$con))
     system2('dropdb', dbname)
